cleanup: move global stats to utils method
