+++ src/pages/staking/Staking.vue
@@ -151,10 +151,8 @@
 </template>
 
 <script>
-import { ethers } from 'ethers';
-import Big from 'big.js';
 import { mapGetters } from 'vuex';
-import { promptAddToMetamask } from 'pages/staking/staking-utils';
+import { fetchStlosApy, promptAddToMetamask } from 'pages/staking/staking-utils';
 import MetaMaskLogo from 'src/assets/metamask-fox.svg'
 import StakeForm from 'pages/staking/StakeForm';
 import UnstakeForm from 'pages/staking/UnstakeForm';
@@ -259,70 +257,25 @@ export default {
     methods: {
         promptAddToMetamask,
         async fetchGlobalStats() {
-            this.stlosTvl = await this.stlosContractInstance.totalAssets()
+            try {
+                this.stlosTvl = await this.stlosContractInstance.totalAssets();
+            } catch ({ message: tvlError }) {
+                console.error(`Failed to fetch sTLOS TVL: ${tvlError}`);
+                this.stlosTvl = null;
+                this.stlosApy = null;
 
-            if (this.stlosTvl.eq('0')) {
-                this.stlosApy = '0';
                 return;
             }
 
-            const rexPoolResponse = await this.$store.$api.getTableRows({
-                code: 'eosio',
-                scope: 'eosio',
-                table: 'rexpool',
-                limit: '1',
-            });
-
-            if (!rexPoolResponse || !rexPoolResponse.rows.length) {
-                console.error('Failed to fetch rexpool');
+            if (this.stlosTvl === null)
                 return;
+
+            try {
+                this.stlosApy = await fetchStlosApy(this.$store.$api, this.stlosTvl);
+            } catch ({ message: apyError }) {
+                console.error(`Failed to fetch sTLOS APY: ${apyError}`);
+                this.stlosApy = null;
             }
-
-            const distConfigResponse = await this.$store.$api.getTableRows({
-                code: 'exrsrv.tf',
-                scope: 'exrsrv.tf',
-                table: 'config',
-                limit: '1',
-            });
-
-            if (!distConfigResponse || !distConfigResponse.rows.length) {
-                console.error('Failed to fetch exrsrv.tf config');
-                return;
-            }
-
-            const payoutsResponse = await this.$store.$api.getTableRows({
-                code: 'exrsrv.tf',
-                scope: 'exrsrv.tf',
-                table: 'payouts',
-                limit: '100',
-            });
-
-            if (!payoutsResponse || !payoutsResponse.rows.length) {
-                console.error('Failed to fetch exrsrv.tf payouts');
-                return;
-            }
-
-            const rexStats = rexPoolResponse.rows[0];
-            const distConfig = distConfigResponse.rows[0];
-            const payoutRow = payoutsResponse.rows.find((row) => row.to === 'eosio.rex');
-
-            const annualPayout = new Big(payoutRow.amount).times(new Big(60 * 60 * 24 * 365).div(payoutRow.interval));
-            const fixedRatio = new Big(distConfig.ratio).div(100);
-            const rexTotal = new Big(rexStats.total_lendable.split(' ')[0]);
-            const stlosTotal = new Big(ethers.utils.formatEther(this.stlosTvl));
-
-            const balanceRatio = rexTotal.eq(0) ? -1 : stlosTotal.times(fixedRatio).div(rexTotal);
-
-            if (balanceRatio.eq(0)) {
-                this.stlosApy = '0';
-                return;
-            }
-
-            const rexPayout = annualPayout.div(balanceRatio.plus(1));
-            const stlosPayout = annualPayout.minus(rexPayout);
-            this.stlosApy = stlosPayout.div(stlosTotal).times(100).toFixed(2);
-            console.log(`Rex APY: ${rexPayout.div(rexTotal).times(100).toFixed(2)}%`)
-            console.log(`sTLOS APY: ${this.stlosApy}%`)
         },
         fetchBalances() {
             if (!this.address) {
@@ -332,6 +285,7 @@ export default {
                 this.totalUnstakedTlosBalance = null;
                 this.stlosValue = null;
                 this.escrowDeposits = [];
+
                 return;
             }
